use riscv64-unknown-elf- toolchain

+++ kernel/framework/kernel.h
@@ -76,7 +76,7 @@ MODULE(vfs) {
 
 MODULE(uproc) {
   void (*init)();
-  int (*mmap)(void *addr, size_t len, int prot, int flags, int fd, off_t offset);
+  int (*mmap)(void *addr, size_t len, int prot, int flags, int fd, size_t offset);
   int (*fork)();
   int (*execve)(const char *path, char *argv[], char *envp[]);
   int (*brk)(void* addr);
